chore(suite-native): split DeviceCompromisedModal components
