+++ suite-native/module-authenticity-checks/package.json
@@ -11,6 +11,7 @@
         "type-check": "yarn g:tsc --build"
     },
     "dependencies": {
+        "@react-navigation/core": "^6.4.10",
         "@react-navigation/native": "6.1.18",
         "@suite-common/wallet-core": "workspace:*",
         "@suite-native/atoms": "workspace:*",
